# Remove nth's linear search overhead in BCF reader